- fixed cpu usage statistic for Linux, now correct values are delivered for the relative values of cpu usage

+++ HomerMonitor/include/ProcessStatistic.h
@@ -50,8 +50,8 @@ struct ThreadStatisticDescriptor{
     int Priority;
     int PriorityBase;
     int ThreadCount;
-    int MemVirtual;
-    int MemPhysical;
+    unsigned long MemVirtual;
+    unsigned long MemPhysical;
 };
 
 typedef std::list<ThreadStatisticDescriptor> ThreadStatisticList;
